ensure the page is within the viewable portion of the frame
when the height is absolutely set, the behavior of flexbox moves some content
out of the view pane, even if the height is a dynamic value. In order to
prevent this, the height is allowed to be dynamic.  This may have other
impacts.

In addition, content is pushed off the sides and bottom of the pain to give
some whitespace.